+++ genvalidity-vector/src/Data/GenValidity/Vector.hs
@@ -17,9 +17,11 @@ instance GenUnchecked v => GenUnchecked (V.Vector v) where
 
 instance GenValid v => GenValid (V.Vector v) where
     genValid = V.fromList <$> genValid
+    shrinkValid = fmap V.fromList . shrinkValid . V.toList
 
-instance GenInvalid v => GenInvalid (V.Vector v) where
+instance (GenUnchecked v, GenInvalid v) => GenInvalid (V.Vector v) where
     genInvalid = V.fromList <$> genInvalid
+    shrinkInvalid = fmap V.fromList . shrinkInvalid . V.toList
 
 instance (SV.Storable e, GenUnchecked e) => GenUnchecked (SV.Vector e) where
     genUnchecked = SV.fromList <$> genUnchecked
@@ -27,3 +29,4 @@ instance (SV.Storable e, GenUnchecked e) => GenUnchecked (SV.Vector e) where
 
 instance (SV.Storable e, GenValid e) => GenValid (SV.Vector e) where
     genValid = SV.fromList <$> genValid
+    shrinkValid = fmap SV.fromList . shrinkValid . SV.toList