+++ genvalidity-text/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+## [0.6.0.0] - 2019-03-06
+
+### Changed
+
+* Removed the 'GenUnchecked' and 'GenInvalid' instance for Text.
+* Poisoned those two instances as well.
+* Removed an 'upTo' from the implementation of 'GenValid Text'.
+  It should generate better sized Text now.
+
 ## [0.5.1.0] - 2018-08-25
 
 ### Added